replace optional chaining on operators with `Optional.map`
Optional chaining is supported on postfix operators (a historical oversight), but not on prefix operators.

Using `Optional.map` brings symmetry between prefix and postfix operators. It's also more readable because it makes it obvious that the operands are optional. The readability improvement is more significant in the case of prefix operators, where there would be 3 consecutive "?"s otherwise.

Co-Authored-By: Tony Allevato <1062427+allevato@users.noreply.github.com>